[1.17] Add a [overwrite_specials] abilities for replace overwrite_specials attribute
The difficulties posed by the use of the overwrite_specials attribute lead me to consider the idea of a capability that would accomplish this function without the limitations posed by the attribute. This way the selection of overwrite abilities is done using the [overwrite_filter] filter or the one_side_only attribute without the presence of the overwrite_specials attribute interfering.