Add documentation to the filter_ability tests
Rename the ones with `_fail` in the name, because I want to reserve `_fail`
for tests which don't expect to result in `TEST_PASS`.

Separate the setup (into an `[event]name=start`) from the attack (during
Alice's turn), so that the test using the `FILTER_ABILITY_TEST` can do
its own setup too.