OUTDATED COPY OF Fix Weapon special filters can be affected by the weapon special leading to infinite recursion
OUTDATED: I did the RAII stuff on top of this version, so using it while making the RAII a draft

When for example we have [damage_type][filter_self][has_attack|filter_weapon]type=pierce applied to one or any pierce type attack and the unit does not have an unaffected pierce attack, the change in type triggers an infinite recursion.

this PR fixes the problem and other because filter of weapon specials or formula.

Fix https://github.com/wesnoth/wesnoth/issues/8940