Assume an ability is present and active if it reaches the recursion limit
Although it passes all of the existing tests, the logic is broken, as
it's assuming that the ability which triggered the recursion limit is
present. However, all it should assume is that there's an ability
which triggered the recursion, not necessarily the same one.

With RECURSION_LIMIT = 4, it doesn't noticably slow the unit tests; I
haven't tested how badly it might affect an AI turn doing lots of
these calculations.

With RECURSION_LIMIT = 10, the damage_type_with_filter_test takes a
significant about of time.  With a limit of 15 and a release build,
that test alone took 7 minutes on my newly-built PC.