add a tag [affect_distant] to abilities coupled with an affect_distant_max_radius option in scenario/era/resources/campaign
like [affect_adjacent] this tag affects other units than the owner of the ability but this time, it concerns units located in a radius defined in the tag.

I had already tried to make one a few years ago but the code was too heavy and I had given up.

the option affect_distant_max_radius in scenario/era/resources/campaign is necessary to avoid checking all the units on the map all the time, for affect_distant to work, it is absolutely necessary to define the maximum radius in the scenario or campaign where the ability in question will be used, otherwise, by default, the tag [affect_distant] will remain inoperative