Combine the affect_adjacent and affect_distant checks into a single function.
Now that [leadership]value=cumulative=yes is working again, I realize that if the same leadership has both [affect_adjacent] and [affect_distant], an adjacent unit that matches the criteria will have its value added twice.

It therefore seems essential to rethink the affect_adjacent check by centering the adjacent hex check on the unit holding the unit and checking if the central unit is at a distance of 1, then if it is facing the correct direction (this avoids modifying the dir values ​​for inversion).

This also avoids rechecking map units for each hex.