Fix abilities_special_tooltips() so that 'name_affected' is displayed the same way as [specials]
For example, if you have a marksman taught to a distant unit and its side isn't the one currently playing its turn, the 'name_affected' that should appear in its distance attack in the ini sidebar disappears. This PR changes this so that it is displayed inactive like a real special marksman.